Added python script to merge files together into releases
Included in this commit is a feature that allows projects to be written on separate documents, and titled in the order that they should be added to releases. A python script can then be run that merges the documents together into a titled release, allowing version control. An image folder has also been added to store the relevant images for the documents